Change the handling of notes in ingest
Unit ingest now prioritises any notes available on the page, intepreting
them as notes before other options like hyperdrive or speed.
Allows for these key terms to be mentioned in the notes without them
being intepreted as something else, and ignored.
Also changes it so that they are now intepreted only if the notes are
exactly 'notes:', case insensitive.